refactor: move api pages to app router
